move coordinator related classes to new package
