+++ kotlin/concepts/authentication/concept-authentication/src/main/java/io/matthewnelson/concept_authentication/coordinator/AuthenticationRequest.kt
@@ -13,7 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.concept_authentication
+package io.matthewnelson.concept_authentication.coordinator
 
 sealed class AuthenticationRequest {
